switch names for charts
