+++ src/components/value-chart/ValueChart.js
@@ -1,110 +1,837 @@
-import React from 'react';
-import Chart from './Chart';
-import ValueText from './ValueText';
-import {
-  // data1,
-  // data2,
-  // data3,
-  data4,
-  dataColored1,
-  dataColored2,
-  dataColored3,
-} from './data';
-import TimespanSelector from './TimespanSelector';
-import { colors } from '../../styles';
+import PropTypes from 'prop-types';
+import React, { Fragment, PureComponent } from 'react';
+import { maxBy, minBy } from 'lodash';
+import Svg, { Path, Circle } from 'react-native-svg';
+import { State } from 'react-native-gesture-handler';
+import Animated, { Easing } from 'react-native-reanimated';
+import Spline from 'cubic-spline';
+import { contains, timing, delay } from 'react-native-redash';
+import { View } from 'react-native';
+import { deviceUtils } from '../../utils';
+import TimestampText from './TimestampText';
+import ActivityIndicator from '../ActivityIndicator';
+import GestureWrapper from './GestureWrapper';
 
-const dataColored = [dataColored1, dataColored2, dataColored3];
-const dataSwitching = [
-  dataColored,
-  [dataColored1, dataColored2],
-  [dataColored2, dataColored3],
-  [data4],
-];
+const AnimatedPath = Animated.createAnimatedComponent(Path);
+const AnimatedCircle = Animated.createAnimatedComponent(Circle);
 
-const colorsArray = [
-  colors.red,
-  colors.grey,
-  colors.green,
-  colors.purple,
-  colors.red,
-  colors.green,
-  colors.red,
-  colors.purple,
-  colors.green,
-  colors.grey,
-  colors.green,
-  colors.purple,
-];
+const {
+  and,
+  or,
+  eq,
+  add,
+  sub,
+  onChange,
+  Value,
+  Clock,
+  block,
+  event,
+  concat,
+  cond,
+  call,
+  set,
+  neq,
+  multiply,
+  greaterOrEq,
+  lessThan,
+  greaterThan,
+  stopClock,
+} = Animated;
 
-let colorIndex = 0;
+let allSegmentDividers = [];
 
-class ValueChart extends React.Component {
+const simplifyChartData = (data, destinatedNumberOfPoints) => {
+  let allSegmentsPoints = [];
+  let colors = [];
+  let lines = [];
+  let dividers = [];
+  let lastPoints = [];
+  let createdLastPoints = [];
+  if (data.segments.length > 0) {
+    for (let i = 0; i < data.segments.length; i++) {
+      allSegmentsPoints = allSegmentsPoints.concat(data.segments[i].points);
+      allSegmentsPoints[allSegmentsPoints.length - 1] = {
+        ...allSegmentsPoints[allSegmentsPoints.length - 1],
+        lastPoint: true,
+      };
+      lastPoints.push(allSegmentsPoints.length - 1);
+      colors.push(data.segments[i].color);
+      lines.push(data.segments[i].line);
+      dividers.push(data.segments[i].renderStartSeparatator);
+    }
+  }
+  if (allSegmentsPoints.length > destinatedNumberOfPoints) {
+    let destMul = allSegmentsPoints.length / destinatedNumberOfPoints;
+    const maxValue = maxBy(allSegmentsPoints, 'y');
+    const minValue = minBy(allSegmentsPoints, 'y');
+
+    const xMul = Math.floor(
+      (allSegmentsPoints[allSegmentsPoints.length - 1].x -
+        allSegmentsPoints[0].x) /
+        allSegmentsPoints.length
+    );
+    let newData = [];
+    newData.push({
+      x: allSegmentsPoints[0].x - xMul * 10,
+      y: allSegmentsPoints[0].y,
+    });
+    newData.push({
+      x: allSegmentsPoints[0].x,
+      y: allSegmentsPoints[0].y,
+    });
+    for (let i = 2; i < destinatedNumberOfPoints - 2; i++) {
+      const indexPlace = i * destMul;
+      const r = indexPlace % 1;
+      const f = Math.floor(indexPlace);
+
+      const firstValue = allSegmentsPoints[f].y * r;
+      const secondValue = allSegmentsPoints[f + 1].y * (1 - r);
+
+      let finalValue;
+      if (firstValue === maxValue) {
+        finalValue = maxValue;
+      } else if (secondValue === minValue) {
+        finalValue = minValue;
+      } else {
+        finalValue = firstValue + secondValue;
+      }
+      if (i * destMul > lastPoints[createdLastPoints.length]) {
+        createdLastPoints.push(newData.length);
+      }
+      newData.push({
+        x: allSegmentsPoints[0].x + i * xMul,
+        y: finalValue,
+      });
+    }
+    newData.push({
+      x: allSegmentsPoints[0].x + destinatedNumberOfPoints * xMul,
+      y: allSegmentsPoints[allSegmentsPoints.length - 1].y,
+    });
+    newData.push({
+      x: allSegmentsPoints[0].x + destinatedNumberOfPoints * xMul + xMul * 10,
+      y: allSegmentsPoints[allSegmentsPoints.length - 1].y,
+    });
+
+    allSegmentDividers = allSegmentDividers.concat(createdLastPoints);
+
+    let returnData = {
+      allPointsForData: allSegmentsPoints,
+      colors,
+      lastPoints: createdLastPoints,
+      lines,
+      points: newData,
+      startSeparatator: dividers,
+    };
+
+    return returnData;
+  }
+};
+
+const hexToRgb = hex => {
+  let result = /^#?([a-f\d]{2})([a-f\d]{2})([a-f\d]{2})$/i.exec(hex);
+  return result
+    ? [
+        parseInt(result[1], 16),
+        parseInt(result[2], 16),
+        parseInt(result[3], 16),
+      ]
+    : null;
+};
+
+const { BEGAN, ACTIVE, CANCELLED, END, FAILED, UNDETERMINED } = State;
+
+const width = deviceUtils.dimensions.width;
+const height = 170;
+const chartPadding = 16;
+
+const flipY = { transform: [{ scaleX: 1 }, { scaleY: -1 }] };
+
+const indexInterval = 10;
+
+const pickImportantPoints = array => {
+  const result = [];
+  result.push(array[0]);
+  let thresholdIndex = indexInterval;
+  for (let i = 1; i < array.length; i++) {
+    if (i === array.length - 1) {
+      result.push(array[i]);
+    } else if (array[i].y === 0 || array[i].y === 200) {
+      result.push(array[i]);
+      thresholdIndex = i + indexInterval;
+    } else if (i === thresholdIndex) {
+      result.push(array[i]);
+      thresholdIndex += indexInterval;
+    }
+  }
+
+  const xs = [];
+  const ys = [];
+  result.forEach(point => {
+    xs.push(point.x);
+    ys.push(point.y);
+  });
+
+  return { xs, ys };
+};
+
+export default class Chart extends PureComponent {
+  propTypes = {
+    data: PropTypes.array,
+    enableSelect: PropTypes.bool,
+    mode: PropTypes.oneOf(['gesture-managed', 'detailed', 'simplified']),
+    stroke: PropTypes.object,
+  };
+
+  static defaultProps = {
+    enableSelect: true,
+    mode: 'gesture-managed',
+    stroke: { detailed: 1.5, simplified: 3 },
+  };
+
   constructor(props) {
     super(props);
 
-    colorIndex = 0;
+    allSegmentDividers = [];
+    this.data = this.props.data.map(data =>
+      simplifyChartData(data, this.props.amountOfPathPoints)
+    );
+
     this.state = {
-      currentChart: 0,
+      chartData: this.data,
+
+      currentData: this.data[0],
+      hideLoadingBar: false,
+      isLoading: true,
+      shouldRenderChart: true,
     };
 
-    this.data = dataSwitching.map((sectionsData, index) => {
-      return {
-        name: index,
-        segments: sectionsData.map((data, i) => {
-          return {
-            color: colorsArray[colorIndex++],
-            line: i * 5,
-            points: data.map(values => {
-              return { x: values.timestamp, y: values.value };
-            }),
-            renderStartSeparatator:
-              colorIndex % 2 != 0
-                ? {
-                    fill: colorsArray[colorIndex],
-                    r: 7,
-                    stroke: 'white',
-                    strokeWidth: colorIndex + 2,
-                  }
-                : undefined,
-          };
-        }),
-      };
-    });
+    this.clock = new Clock();
+    this.clockReversed = new Clock();
+    this.opacityClock = new Clock();
+    this.opacityClockReversed = new Clock();
+    this.loadingClock = new Clock();
+    this.loadingValue = new Value(1);
+    this.gestureState = new Value(UNDETERMINED);
+    this.panGestureState = new Value(UNDETERMINED);
+    this.handle = undefined;
+    this.value = new Value(this.props.mode === 'detailed' ? 0 : 1);
+    this.opacity = new Value(0);
+    this.shouldSpring = new Value(0);
+    this.chartDay = new Value(1);
+    this.chartWeek = new Value(0);
+    this.chartMonth = new Value(0);
+    this.chartYear = new Value(0);
+    this.shouldReactToGestures = new Value(
+      this.props.mode === 'gesture-managed' ? 1 : 0
+    );
+    this.currentChart = new Value(0);
+
+    this.currentInterval = 1;
+
+    this.animatedPath = undefined;
+    this.animatedDividers = undefined;
+
+    this._configUp = {
+      duration: 500,
+      easing: Easing.bezier(0.55, 0.06, 0.45, 0.94),
+      toValue: 1,
+    };
+    this._configDown = {
+      duration: 500,
+      easing: Easing.bezier(0.55, 0.06, 0.45, 0.94),
+      toValue: 0,
+    };
+
+    this.chartsMulti = [
+      this.chartDay,
+      this.chartWeek,
+      this.chartMonth,
+      this.chartYear,
+    ];
+
+    this.onTapGestureEvent = event([
+      {
+        nativeEvent: {
+          state: state =>
+            cond(
+              or(eq(state, State.BEGAN), eq(state, State.END)),
+              set(this.gestureState, state),
+              delay(
+                cond(
+                  or(
+                    neq(state, State.FAILED),
+                    neq(this.panGestureState, State.ACTIVE)
+                  ),
+                  set(this.gestureState, state)
+                ),
+                100
+              )
+            ),
+          x: x =>
+            cond(
+              and(greaterThan(x, 0), lessThan(x, width)),
+              set(this.touchX, x)
+            ),
+        },
+      },
+    ]);
+
+    this.onHandlerStateChange = event([
+      {
+        nativeEvent: {
+          state: state =>
+            block([
+              set(this.panGestureState, state),
+              cond(neq(ACTIVE, state), set(this.gestureState, state)),
+            ]),
+        },
+      },
+    ]);
   }
 
+  componentDidMount = () => {
+    setTimeout(() => {
+      this.reloadChart(0, true);
+    }, 1000);
+  };
+
+  getSnapshotBeforeUpdate() {
+    if (this.currentInterval !== this.props.currentDataSource) {
+      this.currentChart.setValue(this.props.currentDataSource);
+      this.touchX.setValue(deviceUtils.dimensions.width - 1);
+      this.reloadChart(this.props.currentDataSource);
+    }
+  }
+  touchX = new Value(0);
+  lastTouchX = new Value(0);
+
+  onPanGestureEvent = event(
+    [
+      {
+        nativeEvent: {
+          x: x =>
+            cond(
+              and(
+                neq(this.lastTouchX, x),
+                greaterOrEq(x, 0),
+                lessThan(x, width)
+              ),
+              [set(this.touchX, x), set(this.lastTouchX, x)]
+            ),
+        },
+      },
+    ],
+    { useNativeDriver: true }
+  );
+
+  reloadChart = async (currentInterval, isInitial = false) => {
+    if (currentInterval !== this.currentInterval) {
+      if (isInitial) {
+        await this.setState({
+          isLoading: true,
+        });
+
+        const createdSVG = this.createAnimatedPath();
+        this.animatedPath = createdSVG.paths;
+        this.animatedDividers = createdSVG.points;
+      }
+      setTimeout(async () => {
+        Animated.timing(
+          this.chartsMulti[this.currentInterval],
+          this._configDown
+        ).start();
+        Animated.timing(
+          this.chartsMulti[currentInterval],
+          this._configUp
+        ).start();
+        this.currentInterval = currentInterval;
+
+        await this.setState(prevState => ({
+          currentData: prevState.chartData[currentInterval],
+          isLoading: false,
+        }));
+        this.props.onValueUpdate(
+          this.state.chartData[currentInterval].points[
+            this.state.chartData[currentInterval].points.length - 1
+          ].y
+        );
+      });
+    }
+  };
+
+  createSegmentColorsArray = (segments, data) => {
+    let segmentColors = [];
+    let segmentSwitch = [];
+    for (let i = 0; i < segments.length + 1; i++) {
+      segmentColors.push([]);
+      segmentSwitch.push([]);
+    }
+    for (let i = 0; i < data.length; i++) {
+      let dataIndex = 0;
+      for (let j = 0; j < segments.length + 1; j++) {
+        segmentColors[j].push(dataIndex);
+        if (segments[j] == data[i].lastPoints[dataIndex]) {
+          segmentSwitch[j].push(i);
+          dataIndex++;
+        }
+      }
+    }
+    segmentSwitch.pop();
+    return { colors: segmentColors, dividers: segmentSwitch };
+  };
+
+  createAnimatedPath = () => {
+    let sectionEndPoints = [];
+    const { chartData } = this.state;
+    allSegmentDividers.sort(function(a, b) {
+      return a - b;
+    });
+    let segmentsWithDeletedDuplicates = [];
+
+    allSegmentDividers.forEach(function(x) {
+      if (
+        segmentsWithDeletedDuplicates.length == 0 ||
+        segmentsWithDeletedDuplicates.slice(-1)[0] !== x
+      )
+        segmentsWithDeletedDuplicates.push(x);
+    });
+    allSegmentDividers = segmentsWithDeletedDuplicates;
+
+    const segments = this.createSegmentColorsArray(
+      allSegmentDividers,
+      chartData
+    );
+    let splinePoints = [];
+    for (let i = 0; i < chartData.length; i++) {
+      if (chartData[i].points.length > 0) {
+        const maxValue = maxBy(chartData[i].points, 'y');
+        const minValue = minBy(chartData[i].points, 'y');
+
+        const timestampLength =
+          chartData[i].points[chartData[i].points.length - 1].x -
+          chartData[i].points[0].x;
+
+        const xMultiply = width / timestampLength;
+
+        const yMultiply = height / (maxValue.y - minValue.y);
+        const points = chartData[i].points.map(({ x, y }) => ({
+          x: (x - chartData[i].points[0].x) * xMultiply,
+          y: (y - minValue.y) * yMultiply,
+        }));
+
+        const importantPoints = pickImportantPoints(points);
+        const spline = new Spline(importantPoints.xs, importantPoints.ys);
+        splinePoints.push(
+          points
+            .map(({ x, y }) => {
+              return { x, y1: y, y2: spline.at(x) };
+            })
+            .filter(Boolean)
+        );
+      } else {
+        let emptyArray = new Array(this.props.amountOfPathPoints);
+        for (let j = 0; j < emptyArray.length; j++) {
+          emptyArray[j] = { x: 0, y1: 0, y2: 0 };
+        }
+        splinePoints.push(emptyArray);
+      }
+    }
+
+    let chartNode = (multiplyNode, index, chartIndex) =>
+      multiply(
+        multiplyNode,
+        add(
+          splinePoints[chartIndex][index].y1,
+          multiply(
+            this.value,
+            sub(
+              splinePoints[chartIndex][index].y2,
+              splinePoints[chartIndex][index].y1
+            )
+          )
+        )
+      );
+
+    const allNodes = index => {
+      return chartData.map((_, i) => {
+        return chartNode(this.chartsMulti[i], index, i);
+      });
+    };
+
+    let returnPaths = [];
+    let returnPoints = [];
+    for (let i = 0; i <= allSegmentDividers.length; i++) {
+      const animatedPath = concat(
+        'M 0 0',
+        ...splinePoints[0].flatMap(({ x }, index) => {
+          if (i == 0) {
+            if (index <= allSegmentDividers[i]) {
+              return ['L', x, ' ', add(...allNodes(index))];
+            }
+          } else if (i == allSegmentDividers.length) {
+            if (index == allSegmentDividers[i - 1]) {
+              sectionEndPoints.push({
+                index: segments.dividers[i - 1],
+                opacity: this.chartsMulti[segments.dividers[i - 1]],
+                x,
+                y: add(...allNodes(index)),
+              });
+              return ['M', x, ' ', add(...allNodes(index))];
+            }
+            if (index >= allSegmentDividers[i - 1]) {
+              return ['L', x, ' ', add(...allNodes(index))];
+            }
+          } else {
+            if (index == allSegmentDividers[i - 1]) {
+              sectionEndPoints.push({
+                index: segments.dividers[i - 1],
+                opacity: this.chartsMulti[segments.dividers[i - 1]],
+                x,
+                y: add(...allNodes(index)),
+              });
+              return ['M', x, ' ', add(...allNodes(index))];
+            }
+            if (
+              index >= allSegmentDividers[i - 1] &&
+              index <= allSegmentDividers[i]
+            ) {
+              return ['L', x, ' ', add(...allNodes(index))];
+            }
+          }
+        })
+      );
+
+      const colorMatrix = index => {
+        return index < chartData.length - 1
+          ? cond(
+              eq(this.currentChart, index),
+              Animated.color(
+                ...hexToRgb(chartData[index].colors[segments.colors[i][index]])
+              ),
+              colorMatrix(index + 1)
+            )
+          : cond(
+              eq(this.currentChart, index),
+              Animated.color(
+                ...hexToRgb(chartData[index].colors[segments.colors[i][index]])
+              )
+            );
+      };
+
+      const lineMatrix = index => {
+        return index < chartData.length - 1
+          ? cond(
+              eq(this.currentChart, index),
+              chartData[index].lines[segments.colors[i][index]],
+              lineMatrix(index + 1)
+            )
+          : cond(
+              eq(this.currentChart, index),
+              chartData[index].lines[segments.colors[i][index]]
+            );
+      };
+
+      returnPaths.push(
+        <AnimatedPath
+          key={i}
+          id="main-path"
+          fill="none"
+          stroke={colorMatrix(0)}
+          strokeDasharray={lineMatrix(0)}
+          strokeWidth={add(
+            this.props.stroke.detailed,
+            multiply(
+              this.value,
+              sub(this.props.stroke.simplified, this.props.stroke.detailed)
+            )
+          )}
+          strokeLinejoin="round"
+          strokeLinecap="round"
+          d={animatedPath}
+        />
+      );
+    }
+
+    let something = [0, 0, 0, 0];
+    sectionEndPoints.forEach(element => {
+      if (chartData[element.index].startSeparatator[something[element.index]]) {
+        returnPoints.push(
+          <AnimatedCircle
+            cx={element.x}
+            cy={element.y}
+            r={
+              chartData[element.index].startSeparatator[
+                something[element.index]
+              ].r
+            }
+            stroke={
+              chartData[element.index].startSeparatator[
+                something[element.index]
+              ].stroke
+            }
+            strokeWidth={
+              chartData[element.index].startSeparatator[
+                something[element.index]
+              ].strokeWidth
+            }
+            fill={
+              chartData[element.index].startSeparatator[
+                something[element.index]++
+              ].fill
+            }
+            opacity={element.opacity}
+          />
+        );
+      }
+    });
+
+    return { paths: returnPaths, points: returnPoints };
+  };
+
+  checkValueBoundaries = value => {
+    if (Math.abs(value) > width / 2 - 45) {
+      return value > 0 ? value - 45 : value + 45;
+    }
+    return value;
+  };
+
   render() {
-    const change = this.state.currentChart % 2 == 0 ? 20 : -20; // placeholder
+    const { amountOfPathPoints } = this.props;
+    const { currentData } = this.state;
+    let maxValue = 0,
+      minValue = 0,
+      timePeriod = 0,
+      maxValueDistance = 999,
+      minValueDistance = 999;
+
+    if (currentData.points.length > 0) {
+      maxValue = maxBy(currentData.points, 'y');
+      minValue = minBy(currentData.points, 'y');
+
+      timePeriod =
+        currentData.points[currentData.points.length - 1].x -
+        currentData.points[0].x;
+
+      maxValueDistance = this.checkValueBoundaries(
+        ((maxValue.x - currentData.points[0].x) / timePeriod) * width -
+          width / 2
+      );
+      minValueDistance = this.checkValueBoundaries(
+        ((minValue.x - currentData.points[0].x) / timePeriod) * width -
+          width / 2
+      );
+    }
 
     return (
-      <>
-        <ValueText
-          headerText="PRICE"
-          direction={change > 0}
-          change={change.toFixed(2)}
-          ref={component => {
-            this._text = component;
-          }}
+      <Fragment>
+        <GestureWrapper
+          enabled={this.props.enableSelect}
+          onTapGestureEvent={this.onTapGestureEvent}
+          onPanGestureEvent={this.onPanGestureEvent}
+          onHandlerStateChange={this.onHandlerStateChange}
+        >
+          <Animated.View
+            style={{
+              justifyContent: 'flex-start',
+            }}
+          >
+            <Animated.View
+              style={{
+                opacity: this.loadingValue,
+              }}
+            >
+              <TimestampText
+                style={{ transform: [{ translateX: maxValueDistance }] }}
+              >
+                ${Number(maxValue.y).toFixed(2)}
+              </TimestampText>
+            </Animated.View>
+            <View style={{ flexDirection: 'row' }}>
+              <View
+                style={{
+                  height: 200,
+                  width,
+                }}
+              >
+                <Svg
+                  height={width}
+                  width={width + 2}
+                  viewBox={`2 ${height +
+                    chartPadding -
+                    width} ${width} ${width}`}
+                  preserveAspectRatio="none"
+                  style={flipY}
+                >
+                  {this.animatedPath}
+                  {this.animatedDividers}
+                </Svg>
+              </View>
+              <Animated.View
+                style={[
+                  {
+                    backgroundColor: this.props.barColor,
+                    borderRadius: 2,
+                    height: 180,
+                    position: 'absolute',
+                    top: 10,
+                    width: 2,
+                    zIndex: 10,
+                  },
+                  {
+                    opacity: this.opacity,
+                    transform: [
+                      {
+                        translateX: Animated.add(
+                          this.touchX,
+                          new Animated.Value(-1.5)
+                        ),
+                      },
+                    ],
+                  },
+                ]}
+              />
+            </View>
+            <Animated.View
+              style={{
+                opacity: this.loadingValue,
+              }}
+            >
+              <TimestampText
+                style={{ transform: [{ translateX: minValueDistance }] }}
+              >
+                ${Number(minValue.y).toFixed(2)}
+              </TimestampText>
+            </Animated.View>
+          </Animated.View>
+        </GestureWrapper>
+        {this.state.isLoading && (
+          <View
+            style={{
+              alignItems: 'center',
+              backgroundColor: '#ffffffbb',
+              height: 235,
+              justifyContent: 'center',
+              position: 'absolute',
+              top: 100,
+              width: width,
+            }}
+          >
+            <ActivityIndicator />
+          </View>
+        )}
+        <Animated.Code
+          exec={block([
+            cond(
+              or(eq(this.gestureState, ACTIVE), eq(this.gestureState, BEGAN)),
+              set(this.shouldSpring, 1)
+            ),
+            cond(
+              contains([FAILED, CANCELLED, END], this.gestureState),
+              block([
+                set(this.shouldSpring, 0),
+                call([], () => {
+                  this.props.onValueUpdate(
+                    currentData.points[currentData.points.length - 1].y
+                  );
+                }),
+              ]),
+              onChange(
+                this.touchX,
+                call([this.touchX], ([x]) => {
+                  let curX = 0;
+                  if (x < (width / (amountOfPathPoints + 20)) * 10) {
+                    curX = 0;
+                  } else if (
+                    x >
+                    width - (width / (amountOfPathPoints + 20)) * 10
+                  ) {
+                    curX = width - (width / (amountOfPathPoints + 20)) * 10 - 1;
+                  } else {
+                    curX =
+                      x -
+                      (width / (amountOfPathPoints + 20)) * 10 +
+                      (x / width) * (width / (amountOfPathPoints + 20)) * 10;
+                  }
+                  const calculatedIndex = Math.floor(
+                    curX /
+                      ((width - (width / (amountOfPathPoints + 20)) * 10) /
+                        currentData.points.length)
+                  );
+                  this.props.onValueUpdate(
+                    currentData.points[calculatedIndex].y
+                  );
+                })
+              )
+            ),
+            cond(this.shouldReactToGestures, [
+              cond(
+                and(greaterThan(this.value, 0), eq(this.shouldSpring, 1)),
+                block([
+                  stopClock(this.clockReversed),
+                  set(
+                    this.value,
+                    timing({
+                      clock: this.clock,
+                      duration: 350,
+                      easing: Easing.bezier(0.25, 0.46, 0.45, 0.94),
+                      from: this.value,
+                      to: 0,
+                    })
+                  ),
+                ])
+              ),
+              cond(
+                and(lessThan(this.value, 1), eq(this.shouldSpring, 0)),
+                block([
+                  stopClock(this.clock),
+                  set(
+                    this.value,
+                    timing({
+                      clock: this.clockReversed,
+                      duration: 350,
+                      easing: Easing.bezier(0.25, 0.46, 0.45, 0.94),
+                      from: this.value,
+                      to: 1,
+                    })
+                  ),
+                ])
+              ),
+            ]),
+            cond(
+              and(lessThan(this.opacity, 1), eq(this.shouldSpring, 1)),
+              block([
+                set(
+                  this.opacity,
+                  timing({
+                    clock: this.opacityClock,
+                    duration: 500,
+                    easing: Easing.bezier(0.25, 0.46, 0.45, 0.94),
+                    from: this.opacity,
+                    to: 1,
+                  })
+                ),
+                stopClock(this.opacityClockReversed),
+              ])
+            ),
+            cond(
+              and(greaterThan(this.opacity, 0), eq(this.shouldSpring, 0)),
+              block([
+                set(
+                  this.opacity,
+                  timing({
+                    clock: this.opacityClockReversed,
+                    duration: 500,
+                    easing: Easing.bezier(0.25, 0.46, 0.45, 0.94),
+                    from: this.opacity,
+                    to: 0,
+                  })
+                ),
+                stopClock(this.opacityClock),
+              ])
+            ),
+          ])}
         />
-        <Chart
-          mode="gesture-managed" // "gesture-managed" / "detailed" / "simplified"
-          enableSelect // enable checking value in touched point of chart
-          onValueUpdate={value => {
-            this._text.updateValue(value);
-          }}
-          currentDataSource={this.state.currentChart}
-          amountOfPathPoints={150} // amount of points for switch between charts
-          data={this.data}
-          barColor={change > 0 ? colors.chartGreen : colors.red}
-          stroke={{ detailed: 1.5, simplified: 3 }}
-        />
-        <TimespanSelector
-          reloadChart={index => this.setState({ currentChart: index })}
-          color={change > 0 ? colors.chartGreen : colors.red}
-          isLoading={false}
-        />
-      </>
+      </Fragment>
     );
   }
 }
-
-export default ValueChart;
